Refactor: fetchUserPlaylists를 query 안으로 이동
